+++ lib/pages/new_class/student_permissions.dart
@@ -1,0 +1,549 @@
+import 'package:flutter/material.dart';
+import 'package:future_loading_dialog/future_loading_dialog.dart';
+import 'package:get_storage/get_storage.dart';
+import 'package:matrix/matrix.dart';
+import 'package:vrouter/vrouter.dart';
+
+import '../../services/class_services.dart';
+import '../../widgets/matrix.dart';
+import 'package:matrix/matrix.dart' as sdk;
+
+class StudentPermissions extends StatefulWidget {
+  const StudentPermissions({Key? key}) : super(key: key);
+
+  @override
+  State<StudentPermissions> createState() => _StudentPermissionsState();
+}
+
+class _StudentPermissionsState extends State<StudentPermissions> {
+  bool oneToOneChatsWithinClass = false;
+  bool oneToOneChatsWithinExchanges = false;
+  bool createRooms = false;
+  bool createRoomsInExchanges = false;
+  bool createStories = false;
+  bool shareVideos = false;
+  bool sharePhotos = false;
+  bool shareFiles = false;
+  bool shareLocation = false;
+
+  void setOneToOneChatsWithinClass(bool b) => setState(() => oneToOneChatsWithinClass = b);
+  void setOneToOneChatWithinExchanges(bool b) => setState(() => oneToOneChatsWithinExchanges = b);
+  void setCreateRooms(bool b) => setState(() => createRooms = b);
+  void setCreateRoomsInExchange(bool b) => setState(() => createRoomsInExchanges = b);
+  void setCreateStories(bool b) => setState(() => createStories = b);
+  void setShareVideos(bool b) => setState(() => shareVideos = b);
+  void setSharePhotos(bool b) => setState(() => sharePhotos = b);
+  void setShareFiles(bool b) => setState(() => shareFiles = b);
+  void setShareLocation(bool b) => setState(() => shareLocation = b);
+
+
+   submitAction() async {
+     try{
+       String className;
+       String cityName;
+       String countryName;
+       int languageLevel;
+       String schoolName;
+       String disc;
+       String targetLanguage;
+       String sourceLanguage;
+       bool publicGroup;
+       bool openEnrollment;
+       bool openToExchange;
+
+       try{
+         final box = GetStorage();
+         className = box.read("className");
+          cityName = box.read("cityName");
+          countryName = box.read("countryName");
+          languageLevel = box.read("languageLevel");
+         schoolName = box.read("scoolName");
+         disc = box.read("disc");
+         targetLanguage = box.read("targetLanguage");
+         sourceLanguage = box.read("sourceLanage");
+         publicGroup = box.read("publicGroup");
+         openEnrollment = box.read("openEnrollment");
+         openToExchange = box.read("openToExchange");
+       }catch(e){
+         print(e);
+         print("Unable to fetch Data from storage");
+         return;
+       }
+       final matrix = Matrix.of(context);
+       final roomID = await showFutureLoadingDialog(
+         context: context,
+         future: () => matrix.client.createRoom(
+           preset: publicGroup
+               ? sdk.CreateRoomPreset.publicChat
+               : sdk.CreateRoomPreset.privateChat,
+           creationContent: {'type': RoomCreationTypes.mSpace},
+           visibility: publicGroup ? sdk.Visibility.public : null,
+           roomAliasName: publicGroup && className.isNotEmpty
+               ? className.trim()
+               .toLowerCase()
+               .replaceAll(' ', '_')
+               : null,
+           name: className.isNotEmpty
+               ? className
+               : null,
+         ),
+       );
+       if (roomID.result != null) {
+         print(roomID.result);
+         //  print(roomID.result);
+         // class_code = roomID.result!;
+         //
+         // var room = Matrix.of(context).client.getRoomById(class_code);
+         //
+         // print(roomID.result);
+         // if(room !=null){
+         //   // print("working");
+         //   // print(room.canonicalAlias);
+         //   // if(room.canonicalAlias.isEmpty){
+         //   //   print("rerr");
+         //   // }else{
+         //   //   print("work");
+         //   //   class_url = room.canonicalAlias;
+         //   // }
+         //
+         // }else{
+         //   print("test");
+         // }
+
+         await ClassServices.createClass(
+           context: context,
+           isPublic: publicGroup,
+           isSharePhoto: sharePhotos,
+           isCreateRooms: createRooms,
+           isShareFiles: shareFiles,
+           isShareVideo: shareVideos,
+           isCreateStories: createStories,
+           isShareLocation: shareLocation,
+           isOpenExchange: openToExchange,
+           oneToOneChatExchange: oneToOneChatsWithinExchanges,
+           isOpenEnrollment: openEnrollment,
+           isCreateRoomsExchange: createRoomsInExchanges,
+           oneToOneChatClass: oneToOneChatsWithinClass,
+           className: className.toString(),
+           languageLevel: languageLevel,
+           city:  cityName.toString(),
+           roomId: roomID.result!,
+           desc:  disc.toString(),
+           country:  countryName,
+           dominantLanguage:  sourceLanguage,
+           targetLanguage: targetLanguage,
+           schoolName: schoolName.toString(),
+         );
+       }
+       if (roomID == null) {
+         VRouter.of(context).toSegments(['rooms', roomID.result!, 'details']);
+         print(roomID);
+       }
+
+     }catch(e){
+       print(e);
+
+     }
+
+
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    Size size = MediaQuery.of(context).size;
+    return Scaffold(
+      //backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+        appBar: AppBar(
+          backgroundColor: Theme.of(context).backgroundColor,
+          title: Text(
+            "Create a Class",
+            style: TextStyle(
+                color: Theme.of(context).textTheme.bodyText1!.color,
+                fontSize: 14),
+            overflow: TextOverflow.clip,
+            textAlign: TextAlign.center,
+          ),
+          centerTitle: true,
+          elevation: 10,
+          automaticallyImplyLeading: false,
+          leading:  IconButton(
+            icon: const Icon(Icons.arrow_back, color: Colors.black),
+            onPressed: () {
+              context.vRouter.to("/newclass/class_permissions");
+              print(ModalRoute.of(context)!.settings.name);
+              print(Uri.base);
+            },
+          ),
+        ),
+        body: Container(
+          width: size.width,
+          height: size.height,
+          child: SingleChildScrollView(
+            child:Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              children: [
+                SizedBox(
+                  height: size.height * 0.04,
+                ),
+                Container(
+                  margin: EdgeInsets.symmetric(
+                      horizontal: size.width * 0.1, vertical: size.height * 0.02),
+                  width: size.width,
+                  height: 40,
+                  child: Center(
+                    child: Text(
+                      "Student Permissions",
+                      style: TextStyle().copyWith(
+                          color: Theme.of(context).textTheme.bodyText1!.color,
+                          fontSize: 14),
+                      overflow: TextOverflow.clip,
+                      textAlign: TextAlign.center,
+                    ),
+                  ),
+                ),
+                //switch buttons
+                Container(
+                  constraints: const BoxConstraints(
+                    minWidth: 100,
+                    maxWidth: 500,
+                  ),
+                  child: SingleChildScrollView(
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "1-to-1 chats within class",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: oneToOneChatsWithinClass,
+                          onChanged: setOneToOneChatsWithinClass,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "If you allow 1-to-1 chats, your students can send and receive individual chats with other students within your class. Otherwise, they can only chat in rooms.",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "1-to-1 chats within exchanges",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: oneToOneChatsWithinExchanges,
+                          onChanged: setOneToOneChatWithinExchanges,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "If you allow 1-to-1 chats within exchanges, your students can send and receive individual chats with students in exchange-connected classes.",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Create rooms",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: createRooms,
+                          onChanged: setCreateRooms,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                  padding: EdgeInsets.only(right: 5),
+                                  child: Text(
+                                    "Toggle this on to allow students to create rooms within your class",
+                                    style: TextStyle().copyWith(
+                                        color: Theme.of(context)
+                                            .textTheme
+                                            .bodyText1!
+                                            .color,
+                                        fontSize: 14),
+                                    overflow: TextOverflow.clip,
+                                    textAlign: TextAlign.left,
+                                  )),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Create rooms in exchanges",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: createRoomsInExchanges,
+                          onChanged: setCreateRoomsInExchange,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to create rooms within exchanges.",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.left,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Create Stories",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: createStories,
+                          onChanged: setCreateStories,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to create stories.",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Share Video",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: shareVideos,
+                          onChanged:setShareVideos,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to share videos chats/rooms",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Share Photos",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: sharePhotos,
+                          onChanged: setSharePhotos,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to share photos in chats/rooms.",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Share Files",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: shareFiles,
+                          onChanged: setShareFiles,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to share files in chats/rooms",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                        SwitchListTile.adaptive(
+                          title: Text(
+                            "Share Location",
+                            style: TextStyle().copyWith(
+                                color: Theme.of(context).textTheme.bodyText1!.color,
+                                fontSize: 14),
+                            overflow: TextOverflow.clip,
+                          ),
+                          value: shareLocation,
+                          onChanged: setShareLocation,
+                        ),
+                        Row(
+                          children: [
+                            Expanded(
+                              child: Padding(
+                                padding: EdgeInsets.only(right: 5),
+                                child: Text(
+                                  "Toggle this on to allow students to share their location in chats/rooms",
+                                  style: TextStyle().copyWith(
+                                      color:
+                                      Theme.of(context).textTheme.bodyText1!.color,
+                                      fontSize: 14),
+                                  overflow: TextOverflow.clip,
+                                  textAlign: TextAlign.center,
+                                ),
+                              ),
+                            ),
+                          ],
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+                SizedBox(
+                  height: size.height * 0.03,
+                ),
+                Padding(
+                    padding: EdgeInsets.symmetric(
+                      horizontal: size.width * 0.1,
+                      vertical: size.height * 0.02,
+                    ),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.end,
+                      children: [
+                        Expanded(
+                          child: Center(
+                            child: Text(
+                              "3/4",
+                              style: TextStyle().copyWith(
+                                  color: Theme.of(context).textTheme.bodyText1!.color,
+                                  fontSize: 14),
+                              overflow: TextOverflow.clip,
+                              textAlign: TextAlign.center,
+                            ),
+                          ),
+                        ),
+                        InkWell(
+                          onTap: () {
+                        //    context.vRouter.to("/invite_students");
+                            if(ModalRoute.of(context)!.settings.name =="student_permissions"){
+                             submitAction();
+
+                            }else{
+                              //update
+                            }
+
+                           // widget.controller.submitAction();
+                            // setState(() {
+                            //   widget.controller.createClass = 4;
+                            // });
+                          },
+                          child: Container(
+                            width: 50.0,
+                            height: 50.0,
+                            decoration: BoxDecoration(
+                                color: Theme.of(context).colorScheme.onPrimary ==
+                                    Colors.white
+                                    ? Theme.of(context).primaryColor
+                                    : Theme.of(context).colorScheme.onPrimary,
+                                shape: BoxShape.circle,
+                                border: Border.all(
+                                    color: Theme.of(context).colorScheme.onPrimary ==
+                                        Colors.white
+                                        ? Theme.of(context).primaryColorLight
+                                        : Theme.of(context).colorScheme.onPrimary)),
+                            child: const Icon(
+                              Icons.arrow_right_alt,
+                              color: Colors.white,
+                              size: 25,
+                            ),
+                          ),
+                        )
+                      ],
+                    )),
+                SizedBox(
+                  height: size.height * 0.03,
+                ),
+              ],
+            ),
+          ),
+        ));
+  }
+}
